add embroider scenarios to ember-try (#148)
This adds two scenarios. The first will build the dummy app and test suite using embroider in its most-compatible configuration. The second will build the dummy app and test suite using embroider in its most-optimized configuration.

It's helpful for addons to try both, because apps are likely to start with one and migrate toward the other.