Use `require.resolve()` to check if Prism plugin files exist
The old approach implied that the filesystem layout of the `node_modules` folder is flat, but this does not necessarily have to be the case. For example, when using pnpm as a package manager the `prism` package will be resolvable from `ember-prism`, but it will not be at the location that the old approach expects. This commit should fix the problem by relying on the node module resolution algorithm to determine if certain files of the Prism package exist or not.